moved DecodedSignal from canmatrix to own package, with tests.
