+++ tests/test_canmatrix.py
@@ -439,17 +439,7 @@ def test_canid_repr():
     assert can_id.j1939_str == "DA:0xCD PGN:0x1AB00 SA:0x02"
 
 
-# DecodedSignal tests
-def test_decoded_signal_phys_value(some_signal):
-    signal = canmatrix.canmatrix.Signal(factor="0.1", values={10: "Init"})
-    decoded = canmatrix.canmatrix.DecodedSignal(100, signal)
-    assert decoded.phys_value == decimal.Decimal("10")
 
-
-def test_decoded_signal_named_value():
-    signal = canmatrix.canmatrix.Signal(factor="0.1", values={10: "Init"})
-    decoded = canmatrix.canmatrix.DecodedSignal(10, signal)
-    assert decoded.named_value == "Init"
 
 
 def test_Arbitration_id():
