+++ tests/unit/test_DecodedSignal.py
@@ -1,0 +1,56 @@
+# -*- coding: utf-8 -*-
+# Copyright (c) 2013, Eduard Broecker
+# With contributions 2025, Gabriele Omodeo Vanone
+# All rights reserved.
+#
+# Redistribution and use in source and binary forms, with or without
+# modification, are permitted provided that
+# the following conditions are met:
+#
+#    Redistributions of source code must retain the above copyright notice,
+#    this list of conditions and the following disclaimer.
+#    Redistributions in binary form must reproduce the above copyright notice,
+#    this list of conditions and the following disclaimer in the documentation
+#    and/or other materials provided with the distribution.
+#
+# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+# AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+# IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+# ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
+# LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+# CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,   PROCUREMENT OF
+# SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+# INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
+# CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR  OTHERWISE)
+# ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+# POSSIBILITY OF SUCH DAMAGE.
+
+
+import decimal
+
+import pytest
+from builtins import *
+
+import canmatrix.canmatrix
+from canmatrix import Signal
+from canmatrix import DecodedSignal
+
+
+@pytest.fixture
+def some_signal():
+    return Signal(name="speed", size=6, factor=4.7)
+
+
+# DecodedSignal tests
+
+#TODO: why passing some_signal?
+def test_decoded_signal_phys_value(some_signal):
+    signal = Signal(factor="0.1", values={10: "Init"})
+    decoded = DecodedSignal(100, signal)
+    assert decoded.phys_value == decimal.Decimal("10")
+
+
+def test_decoded_signal_named_value():
+    signal = Signal(factor="0.1", values={10: "Init"})
+    decoded = DecodedSignal(10, signal)
+    assert decoded.named_value == "Init"
